test(accordion): update selectors
