+++ packages/components/src/components/bal-accordion/stories/testing.md
@@ -7,19 +7,30 @@ The Baloise Design System provides a collection of custom cypress commands for o
 <!-- START: human documentation -->
 
 ```typescript
-import { byTestId } from '@baloise/design-system-nest-testing'
+import { byTestId, selectors } from '@baloise/design-system-nest-testing'
 
 describe('Accordion', () => {
   const accordion = byTestId('my-accordion') // [data-testid="my-accordion"]
+
   it('should ...', () => {
-    cy.get(accordion).contains('Show more')
+    cy.get(accordion).find(selectors.accordion.header).contains('Show more')
     cy.get(accordion).balAccordionIsClosed()
     cy.get(accordion).click().balAccordionIsOpen()
-    cy.get(page.accordion).contains('Show less')
+    cy.get(accordion).find(selectors.accordion.header).contains('Show less')
+    cy.get(accordion).find(selectors.accordion.content).contains('My Content')
     cy.get(accordion).click().balAccordionIsClosed()
   })
 })
 ```
+
+### Custom Selectors
+
+A list of the custom commands for this specific component.
+
+| Selector            | Element             |
+| ------------------- | ------------------- |
+| `accordion.header`  | The trigger button. |
+| `accordion.content` | Content wrapper.    |
 
 <!-- END: human documentation -->
