+++ packages/components/src/stories/development/implementation/22-testing.stories.mdx
@@ -40,13 +40,18 @@ Now the adjusted commands from the Baloise Design System should be available in 
 Further documentations for each component commands is documented on the components page.
 
 ```typescript
+import { byTestId, selectors } from '@baloise/design-system-nest-testing'
+
 describe('Accordion', () => {
+  const accordion = byTestId('my-accordion') // [data-testid="my-accordion"]
+
   it('should ...', () => {
-    cy.getByTestId('my-accordion').contains('Show more')
-    cy.getByTestId('my-accordion').balAccordionIsClosed()
-    cy.getByTestId('my-accordion').click().balAccordionIsOpen()
-    cy.getByTestId('my-accordion').contains('Show less')
-    cy.getByTestId('my-accordion').click().balAccordionIsClosed()
+    cy.get(accordion).find(selectors.accordion.header).contains('Show more')
+    cy.get(accordion).balAccordionIsClosed()
+    cy.get(accordion).click().balAccordionIsOpen()
+    cy.get(accordion).find(selectors.accordion.header).contains('Show less')
+    cy.get(accordion).find(selectors.accordion.content).contains('My Content')
+    cy.get(accordion).click().balAccordionIsClosed()
   })
 })
 ```
@@ -78,6 +83,21 @@ describe('Button', () => {
 })
 ```
 
+#### Component Selectors (WIP)
+
+For each component we define internally some data-testid selectors.
+
+```typescript
+import { selectors } from '@baloise/design-system-nest-testing'
+
+describe('Accordion', () => {
+  it('should ...', () => {
+    cy.getByTestId('accordion').find(selectors.accordion.header).contains('Show more')
+    cy.getByTestId('accordion').find(selectors.accordion.content).contains('My Content')
+  })
+})
+```
+
 ### Component Commands
 
 #### getByTestId
@@ -86,6 +106,14 @@ Gets the element / component by the attribute `data-testid` and waits until full
 
 ```typescript
 cy.getByTestId('button')
+```
+
+#### disableAnimation
+
+Disables the animation of all the components.
+
+```typescript
+cy.disableAnimation()
 ```
 
 #### getComponent
@@ -125,6 +153,7 @@ cy.platform('desktop')
 ## Links
 
 - [Testing Package](https://github.com/baloise/design-system/blob/master/packages/testing)
+- [Selectors](https://github.com/baloise/design-system/blob/master/packages/testing/src/selectors.ts)
 - [Custom Commands](https://github.com/baloise/design-system/blob/master/packages/testing/src/commands/custom)
 - [Override Commands](https://github.com/baloise/design-system/blob/master/packages/testing/src/commands/overrides)
