+++ packages/testing/src/commands/custom/bal-select.command.ts
@@ -1,4 +1,5 @@
-import { log, selectors, wrapOptions } from '../helpers'
+import { log, wrapOptions } from '../helpers'
+import { selectors } from '../../selectors'
 
 Cypress.Commands.add(
   'balSelectFindOptions',
