+++ packages/testing/src/commands/overrides/blur.command.ts
@@ -1,5 +1,5 @@
+import { selectors } from '../../selectors'
 import {
-  selectors,
   isAccordion,
   isButton,
   isCheckbox,
